backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pyjwt==2.10.1
python-dotenv==1.0.1
google-genai==1.5.0
pydantic==2.10.4
pydantic-settings==2.7.1
